+++ src/main/java/jscover/report/Main.java
@@ -353,8 +353,11 @@ import java.io.File;
 import java.io.IOException;
 import java.util.Properties;
 import java.util.SortedMap;
+import java.util.logging.Logger;
 
 public class Main {
+    private static final Logger logger = Logger.getLogger(Main.class.getName());
+
     public static void main(String[] args) throws IOException {
         new Main().runMain(args);
     }
@@ -407,7 +410,22 @@ public class Main {
     private void mergeReports() {
         File firstReportDir = config.getMergeDirs().get(0);
         ioUtils.copyDir(firstReportDir, config.getMergeDestDir());
+        mergeJSON();
+        mergeOriginalSource();
+    }
 
+    private void mergeOriginalSource() {
+        File srcDir = new File(config.getMergeDestDir(), jscover.Main.reportSrcSubDir);
+        for (int i = 1; i < config.getMergeDirs().size(); i++) {
+            File jsOriginalSrc = new File(config.getMergeDirs().get(i), jscover.Main.reportSrcSubDir);
+            if (jsOriginalSrc.exists())
+                ioUtils.copyDir(jsOriginalSrc, srcDir);
+            else
+                logger.info(String.format("Couldn't find directory '%s' to merge", jsOriginalSrc));
+        }
+    }
+
+    private void mergeJSON() {
         String[] data = new String[config.getMergeDirs().size()];
         for (int i = 0; i < data.length; i++) {
             File dataFile = new File(config.getMergeDirs().get(i), "jscoverage.json");
@@ -417,10 +435,6 @@ public class Main {
         config.getMergeDestDir().mkdirs();
         File mergedJson = new File(config.getMergeDestDir(), "jscoverage.json");
         ioUtils.copy(jsonDataMerger.toJSON(mergedMap), mergedJson);
-
-        File srcDir = new File(config.getMergeDestDir(), jscover.Main.reportSrcSubDir);
-        for (int i = 1; i < config.getMergeDirs().size(); i++)
-            ioUtils.copyDir(new File( config.getMergeDirs().get(i), jscover.Main.reportSrcSubDir), srcDir);
     }
 
     public void generateLCovDataFile() throws IOException {